requirements.txt
Flask==1.1.2
pandas==1.1.3
python-dotenv==0.19.0
Flask-Limiter==1.4
pytest==6.2.4